refactor(categories): implement offline-first repository
