+++ app/src/main/java/com/example/shmr_finance_app_android/data/datasources/categories/CategoriesRemoteDataSourceImpl.kt
@@ -1,18 +1,9 @@
-package com.example.shmr_finance_app_android.data.datasource
+package com.example.shmr_finance_app_android.data.datasources.categories
 
 import com.example.shmr_finance_app_android.data.model.CategoryDTO
 import com.example.shmr_finance_app_android.data.remote.api.FinanceApiService
 import com.example.shmr_finance_app_android.data.remote.mapper.CategoryRemoteMapper
 import javax.inject.Inject
-
-/**
- * Отвечает за получение категорий из remote-источника (API).
- * Определяет контракт для работы с данными категорий без привязки к конкретной реализации.
- */
-interface CategoriesRemoteDataSource {
-    suspend fun getCategoriesByType(isIncome: Boolean): List<CategoryDTO>
-    suspend fun getAllCategories(): List<CategoryDTO>
-}
 
 /**
  * Реализация [CategoriesRemoteDataSource], отвечающая за:
