+++ app/src/main/java/com/example/shmr_finance_app_android/domain/repository/CategoriesRepository.kt
@@ -8,4 +8,5 @@ import com.example.shmr_finance_app_android.domain.model.CategoryDomain
 interface CategoriesRepository {
     suspend fun getCategoriesByType(isIncome: Boolean): Result<List<CategoryDomain>>
     suspend fun getAllCategories(): Result<List<CategoryDomain>>
+    suspend fun syncCategories(): Result<Unit>
 }
